rename polygon-utilities to polygon-utils
